// backend/src/types.ts
export type ArtifactID = string;

export type ArtifactKind = string;

export type ArtifactMetricGeneration = {
  eventSummaryID: EventSummaryID;
  recipeRunID: RecipeRunID;
  generationID: GenerationID;
  modelID: string;
  modelParams?: EvaluationModelParameters | undefined;
  endTimestamp: ISO8601PreciseTimestamp;
  didComplete: boolean;
};

export type ArtifactPathComponent = {
  kind?: (ArtifactKind | null) | undefined;
  id: ArtifactID;
};

export type ArtifactPathPatternComponent =
  | {
      kind: ArtifactKind;
      id?: (ArtifactID | null) | undefined;
    }
  | {
      id: ArtifactID;
    };

export type ArtifactPathPattern = ArtifactPathPatternComponent[];

export type ArtifactPathRecipeTrigger = {
  id: RecipeTriggerID;
  evaluationGroupID: EvaluationGroupID;
  name: string;
  kind: typeof RecipeTriggerKind.artifactPath;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
  artifactPathPattern: ArtifactPathPattern;
};

export type ArtifactPath = ArtifactPathComponent[];
// ...
export type BaseRecipeStep = {
  id: RecipeStepID;
  name: string;
  kind: RecipeStepKind;
  status: RecipeStepStatus;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
  dependencies: RecipeStepID[];
};

export type BaseRecipeTrigger = {
  id: RecipeTriggerID;
  evaluationGroupID: EvaluationGroupID;
  name: string;
  kind: RecipeTriggerKind;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
};
// ...
export type CopyRecipeStep = {
  id: RecipeStepID;
  name: string;
  kind: typeof RecipeStepKind.copy;
  status: RecipeStepStatus;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
  dependencies: RecipeStepID[];
  inputs: RecipeStepMetricInput[];
  outputs: RecipeStepArtifactOutput[];
};
// ...
export type EvaluateRecipeStep = {
  id: RecipeStepID;
  name: string;
  kind: typeof RecipeStepKind.evaluate;
  status: RecipeStepStatus;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
  dependencies: RecipeStepID[];
  inputs: AnyRecipeStepInput[];
  outputs: (BaseRecipeStepOutput & (RecipeStepSnapshotOutput | RecipeStepMetricOutput))[];
  promptTemplate: string;
  userPrompt: string;
  model: RecipeEvaluationModel;
};

export type EvaluationGroupID = UUID;

export type EvaluationModelConfiguration = {
  id: string;
  displayName: string;
  description: string;
  provider: string;
  costMultiplier: string;
  defaultParameters?: EvaluationModelParameters | undefined;
};

export type EvaluationModelParameters = {
  temperature?: number | undefined;
  topP?: number | undefined;
  maxNewTokens?: number | undefined;
};

export type EventSummaryID = UUID;
// ...
export type ISO8601PreciseTimestamp = string | string;
// ...
export type RecipeID = string;

export type RecipeRunID = string;

export type Recipe = {
  id: RecipeID;
  name: string;
  description?: string | undefined;
  isDeleted?: true | undefined;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
  triggers: RecipeTrigger[];
  steps: RecipeStep[];
};
// ...
export const RecipeStepInputKind = {
  constant: "constant",
  artifact: "artifact",
  metric: "metric",
} as const;
export type RecipeStepInputKind = (typeof RecipeStepInputKind)[keyof typeof RecipeStepInputKind];

export const RecipeStepKind = {
  evaluate: "evaluate",
  copy: "copy",
} as const;
export type RecipeStepKind = (typeof RecipeStepKind)[keyof typeof RecipeStepKind];

export type RecipeStepMetricInput = {
  kind: typeof RecipeStepInputKind.metric;
  token: string;
  input: {
    childArtifactPath: ArtifactPath;
    metricID: MetricID;
  };
};

export type RecipeStepMetricOutput = {
  kind: typeof RecipeStepOutputKind.metric;
  key: string;
  output: {
    childArtifactPath: ArtifactPath;
    metricID: MetricID;
    includeEvidence: boolean;
  };
};

export const RecipeStepOutputKind = {
  artifact: "artifact",
  snapshot: "snapshot",
  metric: "metric",
} as const;
export type RecipeStepOutputKind = (typeof RecipeStepOutputKind)[keyof typeof RecipeStepOutputKind];

export type RecipeStep = BaseRecipeStep & (EvaluateRecipeStep | CopyRecipeStep);

export type RecipeStepSnapshotOutput = {
  kind: typeof RecipeStepOutputKind.snapshot;
  key: string;
  output: {
    childArtifactPath: ArtifactPath;
  };
};

export const RecipeStepStatus = {
  enabled: "enabled",
  disabled: "disabled",
  migrated: "migrated",
  hidden: "hidden",
} as const;
export type RecipeStepStatus = (typeof RecipeStepStatus)[keyof typeof RecipeStepStatus];

export type RecipeTriggerID = string;

export const RecipeTriggerKind = {
  artifactPath: "artifactPath",
} as const;
export type RecipeTriggerKind = (typeof RecipeTriggerKind)[keyof typeof RecipeTriggerKind];

export type RecipeTrigger = BaseRecipeTrigger & ArtifactPathRecipeTrigger;
// ...
export type UUID = string;